change 'Sum of score' type from int to float

+++ src/main/java/ir/mahdi/universityservice/controller/QuestionController.java
@@ -44,15 +44,15 @@ public class QuestionController {
         ExamDTO examById = examRestController.getExamById(examId);
 //        List<Question> questions = examQuestionService.findExamQuestionsByExamId(examId);
         List<ExamQuestion> questions = examQuestionService.findAllByExamId(examId);
-        int examScore = sumOfScores(questions);
+        float examScore = sumOfScores(questions);
         model.addAttribute("questions", questions);
         model.addAttribute("exam", examById);
         model.addAttribute("examScore", examScore);
         return "exam-questions";
     }
 
-    private int sumOfScores(List<ExamQuestion> questions) {
-        int score = 0;
+    private float sumOfScores(List<ExamQuestion> questions) {
+        float score = 0;
         for (ExamQuestion question : questions
         ) {
             score += question.getScore();
